Fix/change id int to long

+++ backend/src/main/kotlin/com/infraView/alarm/domain/AlarmPort.kt
@@ -1,8 +1,8 @@
 package com.infraView.alarm.domain
 
 interface AlarmPort {
-    fun getById(id: Int): Alarm?
+    fun getById(id: Long): Alarm?
     fun getAll(): List<Alarm>
     fun save(alarm: Alarm): Alarm
-    fun deleteById(id: Int)
+    fun deleteById(id: Long)
 }
