+++ backend/src/main/kotlin/com/infraView/telemetry/domain/ManageTelemetryUseCase.kt
@@ -2,7 +2,7 @@ package com.infraView.telemetry.domain
 
 interface ManageTelemetryUseCase {
     fun getById(id: Long): Telemetry?
-    fun getByIncidentId(incidentId: Int): List<Telemetry>
+    fun getByIncidentId(incidentId: Long): List<Telemetry>
     fun add(telemetry: Telemetry): Telemetry
     fun delete(id: Long)
 }
